все константы в одном файле
