+++ js/main.js
@@ -1,12 +1,12 @@
-import {changeFilter, createSimilarObject, getPictureFromThumbnails} from './thumbnails.js';
-import {openUploadWindow, closeUploadWindow, setUploadFormSubmit} from './form-upload.js';
+import { createSimilarObject, getPictureFromThumbnails} from './thumbnails.js';
+import {openUploadWindow, closeUploadWindow, setUploadFormSubmit} from './form/form-upload.js';
 import { getData } from './api.js';
-import { showDataErrorMessage } from './messages.js';
+import { showDataErrorMessage } from './form/messages.js';
 import { debounce } from './util.js';
+import {changeFilter} from './filter.js';
 import { DELAY_TIME } from './constants.js';
 
 openUploadWindow ();
-
 
 getData()
   .then((objData) => {
