+++ js/thumbnails.js
@@ -1,45 +1,7 @@
+import { objectList, objectTemplate } from './constants.js';
 import { createBigPicture, openModal } from './modal.js';
-import { shuffleArray } from './util.js';
+import { getFiltredArray } from './filter.js';
 
-
-const objectList = document.querySelector('.pictures');
-const objectTemplate = document.querySelector('#picture').content.querySelector('.picture');
-const imgFilters = document.querySelector('.img-filters');
-const filterForm = imgFilters.querySelector('.img-filters__form');
-const MAX_RANDOM_FILTER_COUNT = 10;
-
-
-let currentFilter = 'filter-default';
-
-const getFiltredArray = (arr) => {
-
-  if(currentFilter === 'filter-discussed') {
-    return arr.slice().sort((a, b) => b.comments.length - a.comments.length);
-  }
-
-  if(currentFilter === 'filter-random') {
-    return shuffleArray(arr.slice()).slice(0, MAX_RANDOM_FILTER_COUNT);
-  }
-  return arr;
-};
-
-const changeFilter = (cb) => {
-  imgFilters.classList.remove('img-filters--inactive');
-
-  filterForm.addEventListener('click', (evt) => {
-    const targetElement = evt.target;
-    if(!targetElement.classList.contains('img-filters__button')) {
-      return;
-    }
-
-    const act = filterForm.querySelector('.img-filters__button--active');
-    act.classList.remove('img-filters__button--active');
-    targetElement.classList.add('img-filters__button--active');
-
-    currentFilter = targetElement.id;
-    cb();
-  });
-};
 
 const createSimilarObject = (data) => {
 
@@ -77,4 +39,4 @@ const getPictureFromThumbnails = (data) => {
   });
 };
 
-export {createSimilarObject, getPictureFromThumbnails,changeFilter};
+export {createSimilarObject, getPictureFromThumbnails,};
